Add test for ```homomorphic_sum``` function. Test passes but due to limitations in CKKS, it is often noisy.
This was the best result I could get by tweeking the parameters. It doesn't seem to be a problem with how it is being used but a limitation with how the slots are allocated or noise aggregating through iteration sum operations.

```
Expected sum: 6.0
Decrypted result sum: 5.679108336874377e+53

Error during test_homomorphic_sum:
Arrays are not almost equal to 1 decimals
 ACTUAL: np.float64(5.679108336874377e+53)
 DESIRED: 6.0
```